add custom error for best error handling

+++ routes/api/user/user.services.ts
@@ -1,8 +1,9 @@
-import bcrypt from "bcrypt";
 import prisma from "../../../prisma"
 import { Role, User } from "@prisma/client";
 import encryptPassword from "../../../Utils/encryptPassword";
 import { UserDto } from "../../../dtos/user.dto";
+import { NotFoundError } from "../../errors/NotFoundError";
+import { AlreadyTakenError } from "../../errors/AlreadyTakenError";
 
 
 export async function findAllUsers(): Promise<Array<User>> {
@@ -19,7 +20,7 @@ export async function createUser(user: UserDto) {
   try {
     const alreadyExist = await prisma.user.findUnique({ where: { email: user.email } });
 
-    if (alreadyExist != undefined) throw new Error("User email already taken");
+    if (alreadyExist) throw new AlreadyTakenError("email");
 
     const hashedPassword = await encryptPassword(user.password);
     const newUser = await prisma.user.create({
@@ -43,7 +44,7 @@ export async function updateUser(userId: number, updatedUser: UserDto): Promise<
     const existingUser = await prisma.user.findUnique({ where: { id: userId } });
 
     if (!existingUser) {
-      throw new Error("User not found");
+      throw new NotFoundError("User not found");
     }
 
     if (updatedUser.password) {
@@ -60,18 +61,24 @@ export async function updateUser(userId: number, updatedUser: UserDto): Promise<
 
     return updatedUserRecord;
   } catch (error) {
+    if (error instanceof NotFoundError) throw new NotFoundError("User");
     throw new Error('Error updating user');
   }
 }
 
 export async function deleteUser(userId: number): Promise<User> {
   try {
+    const existedUser = await prisma.user.findUnique({ where: { id: userId } });
+
+    if (!existedUser) throw new NotFoundError("User")
+
     const deletedUser = await prisma.user.delete({
       where: { id: userId },
     });
 
     return deletedUser;
   } catch (error) {
+    if (error instanceof NotFoundError) throw new NotFoundError("User");
     throw new Error('Error deleting user');
   }
 }
